+++ src/test/java/step3/domain/car/NumberOfCarsTest.java
@@ -2,7 +2,7 @@ package step3.domain.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.domain.exception.NumberOfCarsOutOfBoundException;
+import step3.exception.NumberOfCarsOutOfBoundException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
