+++ src/test/java/step3/domain/strategy/movement/MovementStrategySelectorTest.java
@@ -2,7 +2,7 @@ package step3.domain.strategy.movement;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.domain.exception.IllegalMovementStrategyException;
+import step3.exception.IllegalMovementStrategyException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
